requirements.txt
alembic==1.4.2
appdirs==1.4.3
asgiref==3.2.3
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
awscli==1.18.19
boto==2.49.0
botocore==1.15.19
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
cx-Oracle==7.3.0
distlib==0.3.0
Django==2.1.5
django-crontab==0.6.0
django-scarface==3.2.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
docutils==0.15.2
filelock==3.0.12
Flask==1.1.1
Flask-API==2.0
flask-marshmallow==0.9.0
Flask-Migrate==2.3.0
Flask-SQLAlchemy==2.1
httpie==2.0.0
idna==2.8
importlib-metadata==1.4.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.5
lazy-object-proxy==1.4.3
Mako==1.1.2
MarkupSafe==1.1.1
marshmallow==2.13.6
mccabe==0.6.1
more-itertools==8.1.0
mysqlclient==1.4.6
packaging==20.1
Pillow==7.0.0
pluggy==0.13.1
psycopg2==2.8.5
py==1.8.1
pyasn1==0.4.8
Pygments==2.5.2
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.6
pytest==5.3.4
python-dateutil==2.8.1
python-editor==1.0.4
python2==1.2
pytz==2019.3
PyYAML==5.3
requests==2.22.0
rsa==3.4.2
s3transfer==0.3.3
six==1.14.0
SQLAlchemy==1.2.13
sqlparse==0.3.0
typed-ast==1.4.1
urllib3==1.25.8
virtualenv==20.0.17
virtualenvwrapper-win==1.2.6
wcwidth==0.1.8
Werkzeug==1.0.0
wrapt==1.11.2
zipp==2.1.0
